model now trains, but val loss is still nan (probably because train and val set are same rn

+++ backbones_unet/utils/trainer.py
@@ -120,8 +120,7 @@ class Trainer:
                 # forward pass
                 with torch.cuda.amp.autocast(enabled=self.scaler is not None):
                     preds = self.model(images)
-                    # loss = self.criterion(preds.float(), labels.float())
-                    loss = self.criterion(preds, labels)
+                    loss = self.criterion(preds.float(), labels.float())
                 if not math.isfinite(loss):
                     msg = f"Loss is {loss}, stopping training!"
                     warnings.warn(msg)
